chore: remove package.json from example
