+++ .circleci/config.yml
@@ -1,34 +1,30 @@
-# -------------------------
-#         ALIASES
-# -------------------------
-
+# ALIASES
 aliases:
-# CACHE
-- &restore-gradle-cache
-  keys:
-  - gradle-cache-{{ checksum "android/build.gradle" }}-{{ checksum "example/android/build.gradle" }}-{{ checksum "example/android/app/build.gradle" }}
-- &save-gradle-cache
-  paths:
-  - ~/.gradle
-  key: gradle-cache-{{ checksum "android/build.gradle" }}-{{ checksum "example/android/build.gradle" }}-{{ checksum "example/android/app/build.gradle" }}
+  # CACHE
+  - &restore-gradle-cache
+    keys:
+      - gradle-cache-{{ checksum "android/build.gradle" }}-{{ checksum "example/android/build.gradle" }}-{{ checksum "example/android/app/build.gradle" }}
+  - &save-gradle-cache
+    paths:
+      - ~/.gradle
+    key: gradle-cache-{{ checksum "android/build.gradle" }}-{{ checksum "example/android/build.gradle" }}-{{ checksum "example/android/app/build.gradle" }}
 
-# ANALYSE
-- &eslint
-  name: ESLint Checks
-  command: yarn lint
+  # ANALYSE
+  - &eslint
+    name: ESLint Checks
+    command: yarn lint
 
-- &type-check
-  name: TypeScript Checks
-  command: yarn type-check
+  - &type-check
+    name: TypeScript Checks
+    command: yarn type-check
 
-- &jest
-  name: Jest Unit Tests
-  command: yarn test
+  - &jest
+    name: Jest Unit Tests
+    command: yarn test
 
-- &flow
-  name: Flow Type Check
-  command: yarn flow
-
+  - &flow
+    name: Flow Type Check
+    command: yarn flow
 
 # -------------------------
 #          JOBS
@@ -42,90 +38,88 @@ jobs:
   linux-checkout:
     executor:
       name: rn/linux_js
-      node_version: '10'
+      node_version: "10"
     steps:
-    - attach_workspace:
-        at: .
-    - checkout
-    - rn/yarn_install
-    - persist_to_workspace:
-        root: .
-        paths: .
+      - attach_workspace:
+          at: .
+      - checkout
+      - rn/yarn_install
+      - persist_to_workspace:
+          root: .
+          paths: .
 
   eslint:
     executor:
       name: rn/linux_js
-      node_version: '10'
+      node_version: "10"
     steps:
-    - attach_workspace:
-        at: .
-    - run: *eslint
+      - attach_workspace:
+          at: .
+      - run: *eslint
 
   type-check:
     executor:
       name: rn/linux_js
-      node_version: '10'
+      node_version: "10"
     steps:
-    - attach_workspace:
-        at: .
-    - run: *type-check
-  
+      - attach_workspace:
+          at: .
+      - run: *type-check
+
   jest:
     executor:
       name: rn/linux_js
-      node_version: '10'
+      node_version: "10"
     steps:
-    - attach_workspace:
-        at: .
-    - run: *jest
+      - attach_workspace:
+          at: .
+      - run: *jest
 
   flow:
     executor:
       name: rn/linux_js
-      node_version: '10'
+      node_version: "10"
     steps:
-    - attach_workspace:
-        at: .
-    - run: *flow
-  
+      - attach_workspace:
+          at: .
+      - run: *flow
+
   android-compile:
     executor: rn/linux_android
     steps:
-    - attach_workspace:
-        at: .
-    - restore_cache: *restore-gradle-cache
-    - run:
-        name: Accept Android licences
-        command: |-
-          yes | sdkmanager --licenses || exit 0
-          yes | sdkmanager --update || exit 0
-    - run:
-        name: Build Android Example App and Library
-        command: |-
-          cd example
-          yarn install
-          cd android
-          ./gradlew clean assembleDebug
-    - save_cache: *save-gradle-cache
-  
+      - attach_workspace:
+          at: .
+      - restore_cache: *restore-gradle-cache
+      - run:
+          name: Accept Android licences
+          command: |-
+            yes | sdkmanager --licenses || exit 0
+            yes | sdkmanager --update || exit 0
+      - run:
+          name: Build Android Example App and Library
+          command: |-
+            cd example/android
+            ./gradlew clean assembleDebug
+      - save_cache: *save-gradle-cache
+
   ios-checkout:
     executor: rn/macos
     steps:
-    - checkout
-    - rn/yarn_install
-    - persist_to_workspace:
-        root: .
-        paths: .
+      - checkout
+      - rn/yarn_install
+      - persist_to_workspace:
+          root: .
+          paths: .
 
   ios-compile:
     executor: rn/macos
     steps:
-    - attach_workspace:
-        at: .
-    - run:
-        name: Build example app
-        command: |-
-          react-native run-ios --project-path example/ios
+      - attach_workspace:
+          at: .
+      - run:
+          name: Build example app
+          command: |-
+            react-native run-ios --project-path example/ios
 
 # -------------------------
 #        WORKFLOWS
@@ -134,23 +128,23 @@ workflows:
   version: 2
   Test:
     jobs:
-    - linux-checkout
-    - eslint:
-        requires:
-        - linux-checkout
-    - type-check:
-        requires:
-        - linux-checkout
-    - jest:
-        requires:
-        - linux-checkout
-    - flow:
-        requires:
-        - linux-checkout
-    - android-compile:
-        requires:
-        - linux-checkout
-# Disabled until we have macOS containers enabled        
+      - linux-checkout
+      - eslint:
+          requires:
+            - linux-checkout
+      - type-check:
+          requires:
+            - linux-checkout
+      - jest:
+          requires:
+            - linux-checkout
+      - flow:
+          requires:
+            - linux-checkout
+      - android-compile:
+          requires:
+            - linux-checkout
+# Disabled until we have macOS containers enabled
 #     - ios-checkout
 #     - ios-compile:
 #         requires:
