+++ android/build.gradle
@@ -32,4 +32,3 @@ repositories {
 dependencies {
     implementation 'com.facebook.react:react-native:+'
 }
-
