+++ example/ios/AndroidProgressBarExample.xcodeproj/xcshareddata/xcschemes/AndroidProgressBarExample-tvOS.xcscheme
@@ -17,7 +17,7 @@
                BlueprintIdentifier = "2D2A28121D9B038B00D4039D"
                BuildableName = "libReact.a"
                BlueprintName = "React-tvOS"
-               ReferencedContainer = "container:../../node_modules/react-native/React/React.xcodeproj">
+               ReferencedContainer = "container:../../../node_modules/react-native/React/React.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
