+++ example/ios/AndroidProgressBarExample.xcodeproj/xcshareddata/xcschemes/AndroidProgressBarExample.xcscheme
@@ -17,7 +17,7 @@
                BlueprintIdentifier = "83CBBA2D1A601D0E00E9B192"
                BuildableName = "libReact.a"
                BlueprintName = "React"
-               ReferencedContainer = "container:../../node_modules/react-native/React/React.xcodeproj">
+               ReferencedContainer = "container:../../../node_modules/react-native/React/React.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
