--- example/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "example",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
-    "start": "react-native start",
-    "test": "jest"
-  },
-  "dependencies": {
-    "react": "16.9.0",
-    "react-native": "0.61.5",
-    "react-timer-mixin": "^0.13.4",
-    "@react-native-community/progress-bar-android": "file:../"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.7.5",
-    "@babel/runtime": "^7.7.6",
-    "metro-react-native-babel-preset": "^0.57.0",
-    "react-test-renderer": "16.9.0"
-  }
-}
