+++ js/index.js
@@ -1,0 +1,1 @@
+export {default as ProgressBar} from './RNCProgressBarAndroid';
