+++ package.json
@@ -2,12 +2,11 @@
   "name": "@react-native-community/progress-bar-android",
   "version": "1.0.3",
   "description": "Progress Bar component for React Native",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "./js/index.js",
+  "types": "./js/index.d.ts",
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "start:android": "adb shell am start -n com.androidprogressbar/.MainActivity",
-    "build:android": "cd example/android && ./gradlew clean assembleDebug",
+    "start": "react-native start",
+    "android": "react-native run-android --root example",
     "test": "jest",
     "lint": "eslint 'js/**/*.js' 'example/**/*.js'",
     "type-check": "tsc --noEmit",
@@ -45,8 +44,9 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/runtime": "^7.4.3",
+    "react-timer-mixin": "^0.13.4",
+    "@babel/core": "^7.10.3",
+    "@babel/runtime": "^7.10.3",
     "@react-native-community/eslint-config": "^0.0.7",
     "@types/react-native": "^0.62.2",
     "babel-jest": "^25.1.0",
