+++ react-native.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  project: {
+    android: {sourceDir: './example/android'},
+    ios: {project: './example/ios/example.xcworkspace'},
+  },
+};
